Separado as views das modais
