+++ resources/views/layout/welcome.blade.php
